# backend/api-gateway/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SportBook API",
    "version": "4.0.0",
    "description": "REST API of the SportBook venue booking platform. The API Gateway forwards requests over RabbitMQ RPC to the User, Venue, Booking, Notification and Payment microservices.\n\n**Auth:** call `POST /user/login`, then click **Authorize** and paste the token.\n\n**Payments:** when `STRIPE_SECRET_KEY` is configured, paid slots are booked as `pending_payment` and `POST /bookings/create` returns a `checkout_url` (Stripe Checkout). The slot is held for 30 minutes; the booking becomes `booked` once Stripe confirms the payment (webhook or `GET /payments/verify` on return). Cancelling a paid booking refunds it in full."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Slot is not available"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "avatar_color": {
            "type": "string",
            "nullable": true,
            "example": "#7c3aed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "example": "/api/user/avatar/0b6f…?v=1791276560798"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "description": "Too many attempts"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in and get a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "demo@sportbook.dev"
                  },
                  "password": {
                    "type": "string",
                    "example": "demo1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "description": "Too many attempts"
          }
        }
      }
    },
...
    "/user/avatar": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Upload a profile photo",
        "description": "The client crops and resizes the image (the web app sends a 256×256 WebP).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dataUrl"
                ],
                "properties": {
                  "dataUrl": {
                    "type": "string",
                    "example": "data:image/webp;base64,UklGR…"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "413": {
            "description": "Image larger than 300 KB"
          }
        }
      },
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Remove the profile photo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/user/password/forgot": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Email a password reset link",
        "description": "Always answers 200 so it cannot be used to discover registered emails. Rate limited.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If the email is registered, a link has been sent"
          },
          "429": {
            "description": "Too many attempts"
          }
        }
      }
    },
    "/user/password/reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Set a new password with the emailed token (valid 1 hour, single use)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "newPassword"
                ],
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/payments/webhook": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Stripe webhook endpoint",
        "description": "Called by Stripe, not by the app. The raw body is verified with `Stripe-Signature` and `STRIPE_WEBHOOK_SECRET`; duplicate events are acknowledged without reprocessing.",
        "parameters": [
          {
            "name": "Stripe-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Stripe Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    },
                    "duplicate": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid signature"
          },
          "500": {
            "description": "Processing failed — Stripe retries"
          }
        }
      }
    }
  }
}